#109 Merged framework into other projects
