+++ src/AdminAssistant.sln
@@ -9,8 +9,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Client", "Client", "{0395D6
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Server", "Server", "{6E81FF60-9234-4CBA-B632-640F401017A8}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Framework", "Framework", "{5B790840-9D97-45F2-894F-F5722C56F31F}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AdminAssistant.Blazor.Client", "AdminAssistant.Blazor\Client\AdminAssistant.Blazor.Client.csproj", "{A1B1A4FA-E807-4CAA-A5F1-7380B14042D4}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AdminAssistant.Blazor.Server", "AdminAssistant.Blazor\Server\AdminAssistant.Blazor.Server.csproj", "{32D85134-80F4-49D2-929E-8204E772360C}"
@@ -18,10 +16,6 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AdminAssistant.Test", "AdminAssistant.Test\AdminAssistant.Test.csproj", "{0ADEA26B-F913-4BC2-9860-3E9523C844ED}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AdminAssistant.UI", "AdminAssistant.UI\AdminAssistant.UI.csproj", "{9E39B9F2-0A2E-4423-A4D8-19D6422AC859}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AdminAssistant.Framework", "Framework\AdminAssistant.Framework\AdminAssistant.Framework.csproj", "{344CC4D0-FCD8-484B-A0BC-BDA73A58775A}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AdminAssistant.Framework.Test", "Framework\AdminAssistant.Framework.Test\AdminAssistant.Framework.Test.csproj", "{1ECAFB83-DB49-459F-BED6-07C9D7F9F252}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AdminAssistant", "AdminAssistant\AdminAssistant.csproj", "{9B91A24B-F890-43C7-BB09-F0EA02874480}"
 EndProject
@@ -77,20 +71,6 @@ Global
 		{9E39B9F2-0A2E-4423-A4D8-19D6422AC859}.Release|Any CPU.ActiveCfg = Release|x64
 		{9E39B9F2-0A2E-4423-A4D8-19D6422AC859}.Release|x64.ActiveCfg = Release|x64
 		{9E39B9F2-0A2E-4423-A4D8-19D6422AC859}.Release|x64.Build.0 = Release|x64
-		{344CC4D0-FCD8-484B-A0BC-BDA73A58775A}.Debug|Any CPU.ActiveCfg = Debug|x64
-		{344CC4D0-FCD8-484B-A0BC-BDA73A58775A}.Debug|Any CPU.Build.0 = Debug|x64
-		{344CC4D0-FCD8-484B-A0BC-BDA73A58775A}.Debug|x64.ActiveCfg = Debug|x64
-		{344CC4D0-FCD8-484B-A0BC-BDA73A58775A}.Debug|x64.Build.0 = Debug|x64
-		{344CC4D0-FCD8-484B-A0BC-BDA73A58775A}.Release|Any CPU.ActiveCfg = Release|x64
-		{344CC4D0-FCD8-484B-A0BC-BDA73A58775A}.Release|x64.ActiveCfg = Release|x64
-		{344CC4D0-FCD8-484B-A0BC-BDA73A58775A}.Release|x64.Build.0 = Release|x64
-		{1ECAFB83-DB49-459F-BED6-07C9D7F9F252}.Debug|Any CPU.ActiveCfg = Debug|x64
-		{1ECAFB83-DB49-459F-BED6-07C9D7F9F252}.Debug|Any CPU.Build.0 = Debug|x64
-		{1ECAFB83-DB49-459F-BED6-07C9D7F9F252}.Debug|x64.ActiveCfg = Debug|x64
-		{1ECAFB83-DB49-459F-BED6-07C9D7F9F252}.Debug|x64.Build.0 = Debug|x64
-		{1ECAFB83-DB49-459F-BED6-07C9D7F9F252}.Release|Any CPU.ActiveCfg = Release|x64
-		{1ECAFB83-DB49-459F-BED6-07C9D7F9F252}.Release|x64.ActiveCfg = Release|x64
-		{1ECAFB83-DB49-459F-BED6-07C9D7F9F252}.Release|x64.Build.0 = Release|x64
 		{9B91A24B-F890-43C7-BB09-F0EA02874480}.Debug|Any CPU.ActiveCfg = Debug|x64
 		{9B91A24B-F890-43C7-BB09-F0EA02874480}.Debug|Any CPU.Build.0 = Debug|x64
 		{9B91A24B-F890-43C7-BB09-F0EA02874480}.Debug|x64.ActiveCfg = Debug|x64
@@ -121,8 +101,6 @@ Global
 		{6E81FF60-9234-4CBA-B632-640F401017A8} = {50C8B430-F2F2-4ED1-91A4-8BFB44B9D333}
 		{A1B1A4FA-E807-4CAA-A5F1-7380B14042D4} = {0395D66D-D0ED-4C44-99C0-1CC034CF83D4}
 		{32D85134-80F4-49D2-929E-8204E772360C} = {6E81FF60-9234-4CBA-B632-640F401017A8}
-		{344CC4D0-FCD8-484B-A0BC-BDA73A58775A} = {5B790840-9D97-45F2-894F-F5722C56F31F}
-		{1ECAFB83-DB49-459F-BED6-07C9D7F9F252} = {5B790840-9D97-45F2-894F-F5722C56F31F}
 		{4C7FFB78-374D-4D95-813D-70E2D7F91DED} = {8246FE10-6D6D-4A6E-B3A7-90BA391751FA}
 		{2FA03E41-07AE-4431-863A-20DAD8E4E9C4} = {4C7FFB78-374D-4D95-813D-70E2D7F91DED}
 	EndGlobalSection
