+++ src/AdminAssistant/AdminAssistant.csproj
@@ -28,6 +28,7 @@
 
   <ItemGroup>
     <PackageReference Include="Ardalis.Result" Version="1.0.1" />
+    <PackageReference Include="AutoMapper" Version="10.0.0" />
     <PackageReference Include="FluentValidation" Version="9.0.1" />
     <PackageReference Include="MediatR" Version="8.0.1" />
     <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="8.0.1" />
