+++ src/AdminAssistant/Framework/Providers/LoggingProvider.cs
