+++ src/AdminAssistant/Framework/TypeMapping/IMapFrom.cs
