+++ src/AdminAssistant/Framework/TypeMapping/MappingProfileBase.cs
